test: refactor FileFilters-integ-test
